Makefile: changes for ctags omni completion Keyboard: new keyboard design. TextArea: remove preferred width hack in setSize. TextContainer: make the decoration size publicly accessible.
git-svn-id: https://quirkysoft.googlecode.com/svn/trunk@260 89bda9af-7d2e-0410-a434-4dd01d8ab487

+++ gui-test/arm9/source/UnicodeString.h
@@ -22,7 +22,7 @@
 /** This typedef is used to represent unicode strings. Unicode character
  * values are up to 32 bit so an unsigned int is used.
  */
-typedef std::basic_string<unsigned short> UnicodeString;
+typedef std::basic_string<unsigned int> UnicodeString;
 
 /** Convert a unicode string to a regular char string, by "escaping" unicode characters.
  * @param ustr the unicode string to convert.
